Enhancement/musicplayer (#13)
* 18.05.2021 - 21:28 Music player enhancement.
The old music player was a bit bulky so it got completely reinterpreted and is now added in the MainFrame at the bottom.
Due to the new music player a few bootstrap icons were added and the old background images for the player removed.
Since the app needs a few seconds (depending on the internet connection) to download and extract audio from YouTube, a label indicating the download process has been added.
All existing MouseListeners have been changed from mouseClicked to mouseReleased which results in better compatibility with several mice. (Sometimes clicks were not recognized)
A few additions to ResourceBundle.

* 27.05.2021 - 17:56 Added click on micro icon in player.
A few other GUI related changes.